+++ ReportUtilityBills/tests/utility_bills_test.rb
@@ -395,7 +395,7 @@ class ReportUtilityBillsTest < MiniTest::Test
   def _test_measure(expected_errors: [], expected_warnings: [])
     # Run measure via OSW
     require 'json'
-    template_osw = File.join(File.dirname(__FILE__), '..', '..', 'workflow', 'template.osw')
+    template_osw = File.join(File.dirname(__FILE__), '..', '..', 'workflow', 'template-run-hpxml.osw')
     workflow = OpenStudio::WorkflowJSON.new(template_osw)
     json = JSON.parse(workflow.to_s)