+++ Changelog.md
@@ -1,6 +1,10 @@
 ## OpenStudio-HPXML v1.4.0
+
 __New Features__
 - Updates to OpenStudio 3.4.0/EnergyPlus 22.1.
+- **Breaking change**: New required `OccupancyCalculationType` input to specify operational vs asset calculation. If operational, `NumberofResidents` is required.
+- Expanded capabilities for scheduling:
+  - Allows modeling detailed HVAC setpoints via a schedule CSV file.
 - Allows calculating one or more emissions scenarios (e.g., high renewable penetration vs business as usual) for different emissions types (e.g., CO2e).
 - Allows a heat pump separate backup system to be a central system (e.g., central furnace w/ ducts). Previously only non-central system types were allowed.
 - **Breaking change**: Replaces the `UseMaxLoadForHeatPumps` sizing option with `HeatPumpSizingMethodology`, which has three choices: